create branch hw03-mongodb
add folder controllers.
moved the functions to folder controllers.
checked the work.
