+++ models/contacts.json
@@ -46,11 +46,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "WT5cbe4V9nps_mxW1wKnP",
-    "name": "vlad",
-    "email": "Nazar@gmail.com",
-    "phone": "(073) 423-4906"
   }
 ]
